requirements.txt
markitdown[all]>=0.1.0
pyyaml>=6.0
charset-normalizer>=3.0
magika>=0.5.0
